adjusted bpm range

+++ lib/providers/sample_rate_provider.dart
@@ -1,3 +1,3 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-final countProvider = StateProvider.autoDispose<double>((ref) => 100);
+final countProvider = StateProvider.autoDispose<int>((ref) => 100);
